requirements.txt
absl-py==2.4.0
aeon==1.3.0
astunparse==1.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
cloudpickle==3.1.2
colorama==0.4.6
Deprecated==1.3.1
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.3.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.80.0
h5py==3.14.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
keras==3.12.1
libclang==18.1.1
llvmlite==0.44.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==2.3.3
patsy==1.0.2
protobuf==7.34.1
Pygments==2.20.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyWavelets==1.8.0
PyYAML==6.0.3
requests==2.33.1
rich==14.3.3
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
statsmodels==0.14.6
stumpy==1.14.1
sympy==1.14.0
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
torch==2.11.0
tqdm==4.67.3
tsfresh==0.21.1
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
wrapt==2.1.2
